Solves for issue #6 by adding path string to the context object used on load. Also adds two booleans. One to create the plugin path if not found and another to quiet console messages.
If the provided app path is empty, the default will be used.
If the first character of the path is ~, that will be substitued with the results of os.homedir()
If the last character of the path is :, that will be sibstituted with the App Path per ApplicationPath
Both substitutions are performed with path.join.

README.md updated to match.